+++ projects/plugins/jetpack/changelog/update-move-publicize-to-plugins
@@ -1,0 +1,4 @@
+Significance: minor
+Type: enhancement
+
+Move publicize plugin to the proper extension folder
