+++ projects/plugins/jetpack/extensions/plugins/publicize/store/controls.js
@@ -6,9 +6,9 @@ import apiFetch from '@wordpress/api-fetch';
 /**
  * Trigger an API Fetch request.
  *
- * @param {Object} action Action Object.
- *
- * @return {Promise} Fetch request promise.
+ * @param {Object} action - Action Object.
+ * @param {Object} action.path - Action path.
+ * @returns {Promise} Fetch request promise.
  */
 const fetchFromApi = ( { path } ) => {
 	return apiFetch( { path } );
