+++ projects/plugins/jetpack/extensions/plugins/publicize/store/selectors.js
@@ -9,7 +9,7 @@ import createSelector from 'rememo';
 /**
  * Internal dependencies
  */
-import { SUPPORTED_BLOCKS, SUPPORTED_CONTAINER_BLOCKS } from '../twitter';
+import { SUPPORTED_BLOCKS, SUPPORTED_CONTAINER_BLOCKS } from '../components/twitter';
 
 // Links and media attached to tweets take up 24 characters each.
 const ATTACHMENT_MESSAGE_PADDING = 24;
