+++ tools/eslint-excludelist.json
@@ -108,12 +108,6 @@
 	"projects/plugins/jetpack/extensions/blocks/opentable/use-restaurant-search.js",
 	"projects/plugins/jetpack/extensions/blocks/pinterest/utils.js",
 	"projects/plugins/jetpack/extensions/blocks/podcast-player/utils.js",
-	"projects/plugins/jetpack/extensions/blocks/publicize/connection-verify.js",
-	"projects/plugins/jetpack/extensions/blocks/publicize/form-unwrapped.js",
-	"projects/plugins/jetpack/extensions/blocks/publicize/form.js",
-	"projects/plugins/jetpack/extensions/blocks/publicize/settings-button.js",
-	"projects/plugins/jetpack/extensions/blocks/publicize/store/controls.js",
-	"projects/plugins/jetpack/extensions/blocks/publicize/store/middlewares.js",
 	"projects/plugins/jetpack/extensions/blocks/recurring-payments/index.js",
 	"projects/plugins/jetpack/extensions/blocks/revue/edit.js",
 	"projects/plugins/jetpack/extensions/blocks/slideshow/transforms.js",
